[FIX] pos_hr: barcode and pin code access rights
Since groups added on fields 'barcode' and 'pin' on hr employee in
rev: dfe95e68a1726521423d0a2547c9999b29ded00b

It is not possible anymore to launch the POS when 'pos_hr' module is
installed. Because we don't want to allow users to access in read on all
pin codes and barcodes of employees. We've made a function that will get
a sha1 signature of the barcode and pin code of all employees to store
it in the frontend of POS.

We don't have other solutions concerning the security if we want to keep
the POS working offline. We also know that it is not brute force proof,
as pin code are composed of only digits and you can easily generate the
signatures of all pincodes.

closes odoo/odoo#34447

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>