[FIX] base: Allow to create an user
Now that an onchange set its result into the cache, we
cannot create a user because the onchange tries to put
the reified fields (eg: sel_4_7) into the cache.

As those field don't exist and are pseudo-fields used
to manage groups into the res user form view, we should
replace the values by ORM commands on the groups_id field,
as this is already the case in the create and the write
methods.

Note: Initially merged in saas-12.4, but was not
forward ported, I don't know why.

closes odoo/odoo#34487

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>