[IMP] web: Always display (un)archive buttons in Action on form views
Purpose
=======

While cleaning the stat buttons across all modules, we decided to move
the (un)archiving from stat button to actions in the 'Action' dropdown.

Specification
=============

If the record has an 'active' field, display an action 'Archive' in the
dropdown if the record is active, else display an action 'Unarchive'.

The 'Archive' action should ask to the user if he's sure he wants to
make this action.