[FIX] base: oversight of 87a5f91bf54700198a64c3873044e6568f9a9714
The aforementioned commit prevented the creation of automatic xmlids for
custom fields, however this change is not exactly stable; if a database
already contains automatic xmlids for custom fields and the module to
which these xmlids belong to is upgraded, the xmlids would get deleted,
along with any records related to it because of the way that
`_update_xmlids` works.

This commit leaves existing automatic xmlids of custom fields while
preventing the creation of *new* ones.

closes odoo/odoo#34563

Signed-off-by: Christophe Simonis <chs@odoo.com>