[IMP] hw_drivers: Print Star printers notice
We communicate with receipt printers using the ESCPOS drivers.
By default, Star printers use the Star mode. This can be changed by
modifying the position of the DIP-Switches at the bottom of the
printer.

We provide a link to the documentation to the user if its printer is
not correctly configured.