[FIX] hr_attendance: Allow Marc Demo to checkout
As Marc Demo, at the begining of the month, through
the client action, you can checkin but not checkout.

There is one limitation in the way the attendance state is computed:
It uses the attendance with the most recent checkin date.
Therefore, if an attendance is created in the future, it's the later
attendance which is taken into account, not the current attendance.

Steps to reproduce the problem:

1. Create an attendance in the future with both start and end
date set.

2. Through the client action: checkin

=> the client action stays on the checkin screen because you are
considered as checkout. Hence you cannot checkout.

It's difficult to change the way the state is computed, especially
having the right triggers.

Adding a constraint to avoid attendance in the future is probably a
bad idea. It could unintentionaly fail if the user's time is off by
a few minutes on his computer.

On the other hand, this situation should rarely happen (only when manually
creating an attendance, which has little functional sense).

Solution: Fix demo data to avoid attendances in the future.

Tests needed to be adapted because they used demo data. They no longer
use demo data.