[FIX] mrp: wrong value on cost report
Follow up to:
- commit 1fddc9779d79ccc13261fa27ba3dfae45a9fca09
- commit d247b6960b9af16cb60bcb29ad2f518e891bc31c
- commit 3d1820555a0a82dc30301294d5e577a8cc38dab3

In order to be able to edit a done raw move after the production
and having a correct cost report with consistent values. The
simple rule that a done move should have its quantity equals to
its initial demand. Since report is based on initial demand it
should follow this rule and it's also a good practice because a
lot of computation are based on it (easier than quantity done
because it's already set on the correct quant uom).

opw-2033259

closes odoo/odoo#34695

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>