[IMP] website: hide website specific settings if no website selected
If there is no website selected, there is no point to show the specific
settings (which are supposed to be wrote on website model, that will write on
an empty recordset) as it just flood the screen for nothing.

If multi website is enabled, tell the user he could select a website to load
the settings.
If not enabled, the user can't select a website as the select will be hidden.
In that case, just tell the user that he could create a new website for this
company.

More than just a display improvement, it will also prevent misunderstanding of
the website settings when an user arrive on the settings with no website
selected.

Followup of https://github.com/odoo/odoo/pull/34706

closes odoo/odoo#34764

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>